# Make ImageHDU parametric