This helps make the `size` and `read` functions type-stable.

After this PR:
```julia
julia> fname = tempname() * ".fits"
"/tmp/jl_sFQ6s6.fits"

julia> FITS(fname, "w") do f
           write(f, ones(2,2))
       end

julia> f = FITS(fname);

julia> @code_warntype read(f[1])
Variables
[...]

Body::Array{Float64,2}
[...]

julia> @code_warntype size(f[1])
Variables
[...]

Body::Tuple{Int64,Int64}
[...]
```
As a consequence:
on master:

```julia
julia> FITS(fname, "w") do f
       write(f, ones(100, 100))
       end

julia> f = FITS(fname,"r");

julia> @btime read($(f[1]));
  39.639 μs (12 allocations: 78.89 KiB)
```

After this PR:
```julia
julia> @btime read($(f[1]));
  23.956 μs (5 allocations: 78.48 KiB)
```

The time difference is not significant for large arrays though when reading from disk dominates the execution time.

2. this PR adds a feature to `read!` to read into an array with a different shape, as long as the number of elements match that of the image section to be read in. The signature of `read!` is made a bit stricter, although this is non-breaking (the change is in the type of the error thrown if the `eltype`s don't match).
```julia
julia> FITS(fname, "w") do f
       write(f, ones(2,2))
       end

julia> b = zeros(4);

julia> read!(f[1], b);

julia> b
4-element Array{Float64,1}:
 1.0
 1.0
 1.0
 1.0
```

3. Check if file is opened in read-only mode before writing to it. This used to segfault on master, fixed now. After this PR:
```julia
julia> f = FITS(fname, "r");

julia> write(f, ones(3,3))
ERROR: ArgumentError: FITS file has been opened in read-only mode
```

4. Adds [`Aqua.jl`](https://github.com/JuliaTesting/Aqua.jl) as a test dependency to help maintain project quality.