@jishnub 2021-01-20
Good points, I had missed that. Evidently the only difference in `read` is a reduction in allocations, but that hardly impacts performance. There are marginal performance improvements in `size` and `eltype` and such.

On master:
```julia
julia> fname = tempname() * ".fits";

julia> FITS(fname, "w") do f
       write(f, ones(100, 100))
       end

julia> f = FITS(fname);

julia> sizehdu1(f) = size(f[1]);

julia> eltypehdu1(f) = eltype(f[1]);

julia> @btime sizehdu1($f)
  417.829 ns (2 allocations: 128 bytes)
(100, 100)

julia> @btime eltypehdu1($f)
  7.731 μs (0 allocations: 0 bytes)
Float64
```

After this PR:

```julia
julia> @btime sizehdu1($f)
  159.929 ns (2 allocations: 128 bytes)
(100, 100)

julia> @btime eltypehdu1($f)
  44.430 ns (0 allocations: 0 bytes)
Float64
```

If a typed `ImageHDU` is not something worth adding to this package, I can remove that and retain the other features in this PR.